Create always_block_execution_understanding,v
initially the always@(clk) block execute at t=0s and as the total delay in block in 21s so it will be active till 21s.
so the next triggering activity which will trigger the block will only be reflected in the block after 21s. it will leave 2 clk edge event each time as the always block still active.